@@ -1,13 +1,23 @@
 {
   "name": "react-native-gifted-chat",
-  "version": "0.4.3",
+  "version": "0.7.2",
   "description": "The most complete chat UI for React Native",
   "main": "index.js",
+  "types": "index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/FaridSafi/react-native-gifted-chat.git"
   },
-  "keywords": ["android", "ios", "react-native", "react", "react-component", "messenger", "message", "chat"],
+  "keywords": [
+    "android",
+    "ios",
+    "react-native",
+    "react",
+    "react-component",
+    "messenger",
+    "message",
+    "chat"
+  ],
   "author": "Farid Safi",
   "license": "MIT",
   "bugs": {
@@ -21,19 +31,25 @@
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
     "appr": "cd example-expo && ./setPRVersion.js && yarn && yarn appr",
-    "deploy":
-      "cd example-expo && yarn install && exp login --username $EXP_USERNAME --password $EXP_PASSWORD && exp publish --non-interactive"
+    "deploy": "cd example-expo && yarn install && exp login --username $EXP_USERNAME --password $EXP_PASSWORD && exp publish --non-interactive"
   },
   "jest": {
     "preset": "react-native",
-    "setupFiles": ["./tests/setup.js"],
-    "modulePathIgnorePatterns": ["<rootDir>/example", "<rootDir>/example-expo", "<rootDir>/example-slack-message"]
+    "setupFiles": [
+      "./tests/setup.js"
+    ],
+    "modulePathIgnorePatterns": [
+      "<rootDir>/example",
+      "<rootDir>/example-expo",
+      "<rootDir>/example-slack-message"
+    ]
   },
   "devDependencies": {
     "@types/react": "^16.0.38",
-    "@types/react-native": "^0.52.12",
+    "@types/react-native": "^0.57.35",
     "babel": "6.23.0",
-    "babel-jest": "23.0.0-alpha.0",
+    "babel-eslint": "10.0.1",
+    "babel-jest": "23.0.0",
     "babel-preset-react-native": "4.0.0",
     "eslint": "^4.9.0",
     "eslint-config-airbnb": "16.1.0",
@@ -44,20 +60,23 @@
     "eslint-plugin-react-native": "3.2.1",
     "husky": "0.15.0-rc.2",
     "jest": "22.1.4",
-    "react": "16.2.0",
+    "prop-types": "15.6.2",
+    "react": "16.8.1",
     "react-native": "0.51.0",
-    "react-test-renderer": "16.2.0"
+    "react-test-renderer": "16.8.1"
   },
   "dependencies": {
-    "@expo/react-native-action-sheet": "^1.0.1",
-    "md5": "2.2.1",
+    "@expo/react-native-action-sheet": "^2.0.1",
     "moment": "^2.19.0",
-    "prop-types": "15.6.1",
     "react-native-communications": "2.2.1",
-    "react-native-invertible-scroll-view": "^1.1.0",
     "react-native-lightbox": "^0.7.0",
     "react-native-parsed-text": "^0.0.20",
-    "shallowequal": "1.0.2",
-    "uuid": "3.2.1"
+    "react-native-video": "^3.2.1",
+    "uuid": "3.3.0"
+  },
+  "peerDependencies": {
+    "prop-types": "*",
+    "react": "*",
+    "react-native": "*"
   }
 }
