@@ -1,9 +1,67 @@
-/* eslint-disable */
+import { Linking, MapView } from 'expo';
+import PropTypes from 'prop-types';
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
-import MapView, { PROVIDER_GOOGLE } from 'react-native-maps';
+import { Platform, StyleSheet, TouchableOpacity, ViewPropTypes } from 'react-native';
+
+export default class CustomView extends React.Component {
+
+  static propTypes = {
+    currentMessage: PropTypes.object,
+    containerStyle: ViewPropTypes.style,
+    mapViewStyle: ViewPropTypes.style,
+  };
+
+  static defaultProps = {
+    currentMessage: {},
+    containerStyle: {},
+    mapViewStyle: {},
+  };
+
+  openMapAsync = async () => {
+    const { currentMessage: { location = {} } = {} } = this.props;
+
+    const url = Platform.select({
+      ios: `http://maps.apple.com/?ll=${location.latitude},${location.longitude}`,
+      default: `http://maps.google.com/?q=${location.latitude},${location.longitude}`,
+    });
+
+    try {
+      const supported = await Linking.canOpenURL(url);
+      if (supported) {
+        return Linking.openURL(url);
+      }
+      alert('Opening the map is not supported.');
+    } catch ({ message }) {
+      alert(message);
+    }
+  };
+
+  render() {
+    const { currentMessage, containerStyle, mapViewStyle } = this.props;
+    if (currentMessage.location) {
+      return (
+        <TouchableOpacity style={[styles.container, containerStyle]} onPress={this.openMapAsync}>
+          <MapView
+            style={[styles.mapView, mapViewStyle]}
+            region={{
+              latitude: currentMessage.location.latitude,
+              longitude: currentMessage.location.longitude,
+              latitudeDelta: 0.0922,
+              longitudeDelta: 0.0421,
+            }}
+            scrollEnabled={false}
+            zoomEnabled={false}
+          />
+        </TouchableOpacity>
+      );
+    }
+    return null;
+  }
+
+}
 
 const styles = StyleSheet.create({
+  container: {},
   mapView: {
     width: 150,
     height: 100,
@@ -11,33 +69,3 @@ const styles = StyleSheet.create({
     margin: 3,
   },
 });
-
-export default function CustomView(props) {
-  if (props.currentMessage.location) {
-    return (
-      <View style={props.containerStyle}>
-        <MapView
-          provider={PROVIDER_GOOGLE}
-          style={[styles.mapView]}
-          region={{
-            latitude: props.currentMessage.location.latitude,
-            longitude: props.currentMessage.location.longitude,
-            latitudeDelta: 0.1,
-            longitudeDelta: 0.1,
-          }}
-          scrollEnabled={false}
-          zoomEnabled={false}
-        >
-          <MapView.Marker
-            pinColor={'#fff'}
-            coordinate={{
-              latitude: props.currentMessage.location.latitude,
-              longitude: props.currentMessage.location.longitude,
-            }}
-          />
-        </MapView>
-      </View>
-    );
-  }
-  return null;
-}
