@@ -2,35 +2,27 @@
 
 import PropTypes from 'prop-types';
 import React from 'react';
-import {
-  Text,
-  Clipboard,
-  StyleSheet,
-  TouchableWithoutFeedback,
-  View,
-  ViewPropTypes,
-} from 'react-native';
+import { Text, Clipboard, StyleSheet, TouchableWithoutFeedback, View, ViewPropTypes } from 'react-native';
 
 import MessageText from './MessageText';
 import MessageImage from './MessageImage';
+import MessageVideo from './MessageVideo';
+
 import Time from './Time';
 import Color from './Color';
 
-import { isSameUser, isSameDay, warnDeprecated } from './utils';
+import { isSameUser, isSameDay } from './utils';
 
 export default class Bubble extends React.Component {
 
-  constructor(props) {
-    super(props);
-    this.onLongPress = this.onLongPress.bind(this);
-    this.detectDoubleTap = this.detectDoubleTap.bind(this);
-  }
-
-  onLongPress() {
+  onLongPress = () => {
     if (this.props.onLongPress) {
       this.props.onLongPress(this.context, this.props.currentMessage);
     } else if (this.props.currentMessage.text) {
-      const options = ['Copy Text', 'Cancel'];
+      const options =
+        this.props.optionTitles.length > 0
+          ? this.props.optionTitles.slice(0, 2)
+          : ['Copy Text', 'Cancel'];
       const cancelButtonIndex = options.length - 1;
       this.context.actionSheet().showActionSheetWithOptions(
         {
@@ -48,7 +40,7 @@ export default class Bubble extends React.Component {
         },
       );
     }
-  }
+  };
 
   handleBubbleToNext() {
     if (
@@ -98,6 +90,17 @@ export default class Bubble extends React.Component {
     return null;
   }
 
+  renderMessageVideo() {
+    if (this.props.currentMessage.video) {
+      const { containerStyle, wrapperStyle, ...messageVideoProps } = this.props;
+      if (this.props.renderMessageVideo) {
+        return this.props.renderMessageVideo(messageVideoProps);
+      }
+      return <MessageVideo {...messageVideoProps} />;
+    }
+    return null;
+  }
+
   renderTicks() {
     const { currentMessage } = this.props;
     if (this.props.renderTicks) {
@@ -106,11 +109,12 @@ export default class Bubble extends React.Component {
     if (currentMessage.user._id !== this.props.user._id) {
       return null;
     }
-    if (currentMessage.sent || currentMessage.received) {
+    if (currentMessage.sent || currentMessage.received || currentMessage.pending) {
       return (
         <View style={styles.tickView}>
           {currentMessage.sent && <Text style={[styles.tick, this.props.tickStyle]}>✓</Text>}
           {currentMessage.received && <Text style={[styles.tick, this.props.tickStyle]}>✓</Text>}
+          {currentMessage.pending && <Text style={[styles.tick, this.props.tickStyle]}>🕓</Text>}
         </View>
       );
     }
@@ -128,6 +132,21 @@ export default class Bubble extends React.Component {
     return null;
   }
 
+  renderUsername() {
+    const { currentMessage } = this.props;
+    if (this.props.renderUsernameOnMessage) {
+      if (currentMessage.user._id === this.props.user._id) {
+        return null;
+      }
+      return (
+        <View style={styles.usernameView}>
+          <Text style={[styles.username, this.props.usernameStyle]}>~ {currentMessage.user.name}</Text>
+        </View>
+      );
+    }
+    return null;
+  }
+
   renderCustomView() {
     if (this.props.renderCustomView) {
       return this.props.renderCustomView(this.props);
@@ -135,7 +154,7 @@ export default class Bubble extends React.Component {
     return null;
   }
 
-  detectDoubleTap () {
+  detectDoubleTap = () => {
     const now = new Date().getTime()
     if (this.lastTap && (now - this.lastTap) < 300) {
       delete this.lastTap
@@ -150,12 +169,7 @@ export default class Bubble extends React.Component {
 
   render() {
     return (
-      <View
-        style={[
-          styles[this.props.position].container,
-          this.props.containerStyle[this.props.position],
-        ]}
-      >
+      <View style={[styles[this.props.position].container, this.props.containerStyle[this.props.position]]}>
         <View
           style={[
             styles[this.props.position].wrapper,
@@ -173,8 +187,10 @@ export default class Bubble extends React.Component {
             <View>
               {this.renderCustomView()}
               {this.renderMessageImage()}
+              {this.renderMessageVideo()}
               {this.renderMessageText()}
-              <View style={[styles.bottom, this.props.bottomContainerStyle[this.props.position]]}>
+              <View style={[styles[this.props.position].bottom, this.props.bottomContainerStyle[this.props.position]]}>
+                {this.renderUsername()}
                 {this.renderTime()}
                 {this.renderTicks()}
               </View>
@@ -206,6 +222,10 @@ const styles = {
     containerToPrevious: {
       borderTopLeftRadius: 3,
     },
+    bottom: {
+      flexDirection: 'row',
+      justifyContent: 'flex-start',
+    },
   }),
   right: StyleSheet.create({
     container: {
@@ -225,11 +245,11 @@ const styles = {
     containerToPrevious: {
       borderTopRightRadius: 3,
     },
+    bottom: {
+      flexDirection: 'row',
+      justifyContent: 'flex-end',
+    },
   }),
-  bottom: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-  },
   tick: {
     fontSize: 10,
     backgroundColor: Color.backgroundTransparent,
@@ -238,6 +258,17 @@ const styles = {
   tickView: {
     flexDirection: 'row',
     marginRight: 10,
+  },
+  username: {
+    top: -3,
+    left: 0,
+    fontSize: 12,
+    backgroundColor: 'transparent',
+    color: '#aaa',
+  },
+  usernameView: {
+    flexDirection: 'row',
+    marginHorizontal: 10,
   },
 };
 
@@ -250,11 +281,14 @@ Bubble.defaultProps = {
   onLongPress: null,
   onDoubleTap: null,
   renderMessageImage: null,
+  renderMessageVideo: null,
   renderMessageText: null,
   renderCustomView: null,
+  renderUsername: null,
   renderTicks: null,
   renderTime: null,
   position: 'left',
+  optionTitles: ['Copy Text', 'Cancel'],
   currentMessage: {
     text: null,
     createdAt: null,
@@ -266,11 +300,9 @@ Bubble.defaultProps = {
   wrapperStyle: {},
   bottomContainerStyle: {},
   tickStyle: {},
+  usernameStyle: {},
   containerToNextStyle: {},
   containerToPreviousStyle: {},
-  // TODO: remove in next major release
-  isSameDay: warnDeprecated(isSameDay),
-  isSameUser: warnDeprecated(isSameUser),
 };
 
 Bubble.propTypes = {
@@ -279,11 +311,15 @@ Bubble.propTypes = {
   onLongPress: PropTypes.func,
   onDoubleTap: PropTypes.func,
   renderMessageImage: PropTypes.func,
+  renderMessageVideo: PropTypes.func,
   renderMessageText: PropTypes.func,
   renderCustomView: PropTypes.func,
+  renderUsernameOnMessage: PropTypes.bool,
+  renderUsername: PropTypes.func,
   renderTime: PropTypes.func,
   renderTicks: PropTypes.func,
   position: PropTypes.oneOf(['left', 'right']),
+  optionTitles: PropTypes.arrayOf(PropTypes.string),
   currentMessage: PropTypes.object,
   nextMessage: PropTypes.object,
   previousMessage: PropTypes.object,
@@ -300,6 +336,7 @@ Bubble.propTypes = {
     right: ViewPropTypes.style,
   }),
   tickStyle: Text.propTypes.style,
+  usernameStyle: Text.propTypes.style,
   containerToNextStyle: PropTypes.shape({
     left: ViewPropTypes.style,
     right: ViewPropTypes.style,
@@ -308,7 +345,4 @@ Bubble.propTypes = {
     left: ViewPropTypes.style,
     right: ViewPropTypes.style,
   }),
-  // TODO: remove in next major release
-  isSameDay: PropTypes.func,
-  isSameUser: PropTypes.func,
 };
