@@ -7,14 +7,14 @@ import moment from 'moment';
 
 import Color from './Color';
 
-import { isSameDay, isSameUser, warnDeprecated } from './utils';
+import { isSameDay } from './utils';
 import { DATE_FORMAT } from './Constant';
 
 export default function Day(
-  { dateFormat, currentMessage, previousMessage, containerStyle, wrapperStyle, textStyle },
+  { dateFormat, currentMessage, previousMessage, nextMessage, containerStyle, wrapperStyle, textStyle, inverted },
   context,
 ) {
-  if (!isSameDay(currentMessage, previousMessage)) {
+  if (!isSameDay(currentMessage, inverted ? previousMessage : nextMessage)) {
     return (
       <View style={[styles.container, containerStyle]}>
         <View style={wrapperStyle}>
@@ -56,23 +56,20 @@ Day.defaultProps = {
     createdAt: null,
   },
   previousMessage: {},
+  nextMessage: {},
   containerStyle: {},
   wrapperStyle: {},
   textStyle: {},
-  // TODO: remove in next major release
-  isSameDay: warnDeprecated(isSameDay),
-  isSameUser: warnDeprecated(isSameUser),
   dateFormat: DATE_FORMAT,
 };
 
 Day.propTypes = {
   currentMessage: PropTypes.object,
   previousMessage: PropTypes.object,
+  nextMessage: PropTypes.object,
+  inverted: PropTypes.bool,
   containerStyle: ViewPropTypes.style,
   wrapperStyle: ViewPropTypes.style,
   textStyle: Text.propTypes.style,
-  // TODO: remove in next major release
-  isSameDay: PropTypes.func,
-  isSameUser: PropTypes.func,
   dateFormat: PropTypes.string,
 };
