@@ -1,4 +1,4 @@
-/* eslint no-use-before-define: ["error", { "variables": false }], react-native/no-inline-styles: 0 */
+/* eslint react-native/no-inline-styles: 0 */
 
 import PropTypes from 'prop-types';
 import React from 'react';
@@ -11,16 +11,55 @@ import Day from './Day';
 
 import { isSameUser, isSameDay } from './utils';
 
+const styles = {
+  left: StyleSheet.create({
+    container: {
+      flexDirection: 'row',
+      alignItems: 'flex-end',
+      justifyContent: 'flex-start',
+      marginLeft: 8,
+      marginRight: 0,
+    },
+  }),
+  right: StyleSheet.create({
+    container: {
+      flexDirection: 'row',
+      alignItems: 'flex-end',
+      justifyContent: 'flex-end',
+      marginLeft: 0,
+      marginRight: 8,
+    },
+  }),
+};
+
 export default class Message extends React.Component {
 
-  getInnerComponentProps() {
+  shouldComponentUpdate(nextProps) {
+    const next = nextProps.currentMessage;
+    const current = this.props.currentMessage;
+    const { nextMessage } = this.props;
+    const nextPropsMessage = nextProps.nextMessage;
+    return (
+      next.sent !== current.sent ||
+      next.received !== current.received ||
+      next.pending !== current.pending ||
+      next.createdAt !== current.createdAt ||
+      next.text !== current.text ||
+      next.image !== current.image ||
+      next.video !== current.video ||
+      next.audio !== current.audio ||
+      nextMessage !== nextPropsMessage
+    );
+  }
+
+  getInnerComponentProps = () => {
     const { containerStyle, ...props } = this.props;
     return {
       ...props,
       isSameUser,
       isSameDay,
     };
-  }
+  };
 
   renderDay() {
     if (this.props.currentMessage.createdAt) {
@@ -87,27 +126,6 @@ export default class Message extends React.Component {
   }
 
 }
-
-const styles = {
-  left: StyleSheet.create({
-    container: {
-      flexDirection: 'row',
-      alignItems: 'flex-end',
-      justifyContent: 'flex-start',
-      marginLeft: 8,
-      marginRight: 0,
-    },
-  }),
-  right: StyleSheet.create({
-    container: {
-      flexDirection: 'row',
-      alignItems: 'flex-end',
-      justifyContent: 'flex-end',
-      marginLeft: 0,
-      marginRight: 8,
-    },
-  }),
-};
 
 Message.defaultProps = {
   renderAvatar: undefined,
