@@ -3,86 +3,70 @@
     no-param-reassign: 0,
     no-use-before-define: ["error", { "variables": false }],
     no-return-assign: 0,
-    react/no-string-refs: 0
+    react/no-string-refs: 0,
+    react/sort-comp: 0
 */
 
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { ListView, View, StyleSheet } from 'react-native';
+import { FlatList, View, StyleSheet, Keyboard, TouchableOpacity, Text } from 'react-native';
 
-import shallowequal from 'shallowequal';
-import InvertibleScrollView from 'react-native-invertible-scroll-view';
-import md5 from 'md5';
 import LoadEarlier from './LoadEarlier';
 import Message from './Message';
+import Color from './Color';
 
-export default class MessageContainer extends React.Component {
+export default class MessageContainer extends React.PureComponent {
 
-  constructor(props) {
-    super(props);
+  state = {
+    showScrollBottom: false,
+  };
 
-    this.renderRow = this.renderRow.bind(this);
-    this.renderFooter = this.renderFooter.bind(this);
-    this.renderLoadEarlier = this.renderLoadEarlier.bind(this);
-    this.renderScrollComponent = this.renderScrollComponent.bind(this);
+  componentDidMount() {
+    if (this.props.messages.length === 0) {
+      this.attachKeyboardListeners();
+    }
+  }
 
-    const dataSource = new ListView.DataSource({
-      rowHasChanged: (r1, r2) => {
-        return r1.hash !== r2.hash;
-      },
-    });
-
-    const messagesData = this.prepareMessages(props.messages);
-    this.state = {
-      dataSource: dataSource.cloneWithRows(messagesData.blob, messagesData.keys),
-    };
+  componentWillUnmount() {
+    this.detachKeyboardListeners();
   }
 
   componentWillReceiveProps(nextProps) {
-    if (this.props.messages === nextProps.messages) {
-      return;
+    if (this.props.messages.length === 0 && nextProps.messages.length > 0) {
+      this.detachKeyboardListeners();
+    } else if (this.props.messages.length > 0 && nextProps.messages.length === 0) {
+      this.attachKeyboardListeners(nextProps);
     }
-    const messagesData = this.prepareMessages(nextProps.messages);
-    this.setState({
-      dataSource: this.state.dataSource.cloneWithRows(messagesData.blob, messagesData.keys),
-    });
   }
 
-  shouldComponentUpdate(nextProps, nextState) {
-    if (!shallowequal(this.props, nextProps)) {
-      return true;
+  attachKeyboardListeners = () => {
+    const { invertibleScrollViewProps: invertibleProps } = this.props;
+    Keyboard.addListener('keyboardWillShow', invertibleProps.onKeyboardWillShow);
+    Keyboard.addListener('keyboardDidShow', invertibleProps.onKeyboardDidShow);
+    Keyboard.addListener('keyboardWillHide', invertibleProps.onKeyboardWillHide);
+    Keyboard.addListener('keyboardDidHide', invertibleProps.onKeyboardDidHide);
+  };
+
+  detachKeyboardListeners = () => {
+    const { invertibleScrollViewProps: invertibleProps } = this.props;
+    Keyboard.removeListener('keyboardWillShow', invertibleProps.onKeyboardWillShow);
+    Keyboard.removeListener('keyboardDidShow', invertibleProps.onKeyboardDidShow);
+    Keyboard.removeListener('keyboardWillHide', invertibleProps.onKeyboardWillHide);
+    Keyboard.removeListener('keyboardDidHide', invertibleProps.onKeyboardDidHide);
+  };
+
+  renderFooter = () => {
+    if (this.props.renderFooter) {
+      const footerProps = {
+        ...this.props,
+      };
+      return this.props.renderFooter(footerProps);
     }
-    if (!shallowequal(this.state, nextState)) {
-      return true;
-    }
-    return false;
-  }
+    return null;
+  };
 
-  prepareMessages(messages) {
-    return {
-      keys: messages.map((m) => m._id),
-      blob: messages.reduce((o, m, i) => {
-        const previousMessage = messages[i + 1] || {};
-        const nextMessage = messages[i - 1] || {};
-        // add next and previous messages to hash to ensure updates
-        const toHash = JSON.stringify(m) + previousMessage._id + nextMessage._id;
-        o[m._id] = {
-          ...m,
-          previousMessage,
-          nextMessage,
-          hash: md5(toHash),
-        };
-        return o;
-      }, {}),
-    };
-  }
-
-  scrollTo(options) {
-    this._invertibleScrollViewRef.scrollTo(options);
-  }
-
-  renderLoadEarlier() {
+  renderLoadEarlier = () => {
     if (this.props.loadEarlier === true) {
       const loadEarlierProps = {
         ...this.props,
@@ -93,78 +77,102 @@ export default class MessageContainer extends React.Component {
       return <LoadEarlier {...loadEarlierProps} />;
     }
     return null;
+  };
+
+  scrollTo(options) {
+    if (this.flatListRef && options) {
+      this.flatListRef.scrollToOffset(options);
+    }
   }
 
-  renderFooter() {
-    if (this.props.renderFooter) {
-      const footerProps = {
-        ...this.props,
-      };
-      return this.props.renderFooter(footerProps);
-    }
-    return null;
-  }
+  scrollToBottom = () => {
+    this.scrollTo({ offset: 0, animated: 'true' });
+  };
 
-  renderRow(message) {
-    if (!message._id && message._id !== 0) {
-      console.warn('GiftedChat: `_id` is missing for message', JSON.stringify(message));
+  handleOnScroll = (event) => {
+    if (event.nativeEvent.contentOffset.y > this.props.scrollToBottomOffset) {
+      this.setState({ showScrollBottom: true });
+    } else {
+      this.setState({ showScrollBottom: false });
     }
-    if (!message.user) {
-      if (!message.system) {
-        console.warn('GiftedChat: `user` is missing for message', JSON.stringify(message));
+  };
+
+  renderRow = ({ item, index }) => {
+    if (!item._id && item._id !== 0) {
+      console.warn('GiftedChat: `_id` is missing for message', JSON.stringify(item));
+    }
+    if (!item.user) {
+      if (!item.system) {
+        console.warn('GiftedChat: `user` is missing for message', JSON.stringify(item));
       }
-      message.user = {};
+      item.user = {};
     }
+    const { messages, ...restProps } = this.props;
+    const previousMessage = messages[index + 1] || {};
+    const nextMessage = messages[index - 1] || {};
 
     const messageProps = {
-      ...this.props,
-      key: message._id,
-      currentMessage: message,
-      previousMessage: message.previousMessage,
-      nextMessage: message.nextMessage,
-      position: message.user._id === this.props.user._id ? 'right' : 'left',
+      ...restProps,
+      key: item._id,
+      currentMessage: item,
+      previousMessage,
+      nextMessage,
+      position: item.user._id === this.props.user._id ? 'right' : 'left',
     };
 
     if (this.props.renderMessage) {
       return this.props.renderMessage(messageProps);
     }
     return <Message {...messageProps} />;
+  };
+
+  renderHeaderWrapper = () => <View style={styles.headerWrapper}>{this.renderLoadEarlier()}</View>;
+
+  renderScrollToBottomWrapper() {
+    const scrollToBottomComponent = (
+      <View style={styles.scrollToBottomStyle}>
+        <TouchableOpacity onPress={this.scrollToBottom} hitSlop={{ top: 5, left: 5, right: 5, bottom: 5 }}>
+          <Text>V</Text>
+        </TouchableOpacity>
+      </View>
+    );
+
+    if (this.props.scrollToBottomComponent) {
+      return (
+        <TouchableOpacity onPress={this.scrollToBottom} hitSlop={{ top: 5, left: 5, right: 5, bottom: 5 }}>
+          {this.props.scrollToBottomComponent}
+        </TouchableOpacity>
+      );
+    }
+    return scrollToBottomComponent;
   }
 
-  renderScrollComponent(props) {
-    const { invertibleScrollViewProps } = this.props;
-    return (
-      <InvertibleScrollView
-        {...props}
-        {...invertibleScrollViewProps}
-        ref={(component) => (this._invertibleScrollViewRef = component)}
-      />
-    );
-  }
+  keyExtractor = (item) => `${item._id}`;
 
   render() {
-    const contentContainerStyle = this.props.inverted
-      ? {}
-      : styles.notInvertedContentContainerStyle;
-
-    if (this.props.inverted && this.props.topOffset) {
-      contentContainerStyle.paddingBottom = this.props.topOffset
+    if (this.props.messages.length === 0) {
+      return <View style={styles.container} />;
     }
-
     return (
-      <View style={styles.container}>
-        <ListView
+      <View style={this.props.alignTop ? styles.containerAlignTop : styles.container}>
+        {this.state.showScrollBottom && this.props.scrollToBottom ? this.renderScrollToBottomWrapper() : null}
+        <FlatList
+          ref={(ref) => (this.flatListRef = ref)}
+          extraData={this.props.extraData}
+          keyExtractor={this.keyExtractor}
           enableEmptySections
           automaticallyAdjustContentInsets={false}
-          initialListSize={20}
-          pageSize={20}
+          inverted={this.props.inverted}
+          data={this.props.messages}
+          style={styles.listStyle}
+          contentContainerStyle={styles.contentContainerStyle}
+          renderItem={this.renderRow}
+          {...this.props.invertibleScrollViewProps}
+          ListFooterComponent={this.renderHeaderWrapper}
+          ListHeaderComponent={this.renderFooter}
+          onScroll={this.handleOnScroll}
+          scrollEventThrottle={100}
           {...this.props.listViewProps}
-          dataSource={this.state.dataSource}
-          contentContainerStyle={contentContainerStyle}
-          renderRow={this.renderRow}
-          renderHeader={this.props.inverted ? this.renderFooter : this.renderLoadEarlier}
-          renderFooter={this.props.inverted ? this.renderLoadEarlier : this.renderFooter}
-          renderScrollComponent={this.renderScrollComponent}
         />
       </View>
     );
@@ -176,8 +184,37 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
   },
-  notInvertedContentContainerStyle: {
+  containerAlignTop: {
+    flexDirection: 'row',
+    alignItems: 'flex-start',
+  },
+  contentContainerStyle: {
     justifyContent: 'flex-end',
+  },
+  headerWrapper: {
+    flex: 1,
+  },
+  listStyle: {
+    flex: 1,
+  },
+  scrollToBottomStyle: {
+    opacity: 0.8,
+    position: 'absolute',
+    paddingHorizontal: 15,
+    paddingVertical: 8,
+    right: 10,
+    bottom: 30,
+    zIndex: 999,
+    height: 40,
+    width: 40,
+    borderRadius: 20,
+    backgroundColor: Color.white,
+    alignItems: 'center',
+    justifyContent: 'center',
+    shadowColor: Color.black,
+    shadowOpacity: 0.5,
+    shadowOffset: { width: 0, height: 0 },
+    shadowRadius: 1,
   },
 });
 
@@ -186,11 +223,15 @@ MessageContainer.defaultProps = {
   user: {},
   renderFooter: null,
   renderMessage: null,
-  onLoadEarlier: () => { },
+  onLoadEarlier: () => {},
   inverted: true,
   loadEarlier: false,
   listViewProps: {},
   invertibleScrollViewProps: {},
+  extraData: null,
+  scrollToBottom: false,
+  scrollToBottomOffset: 200,
+  alignTop: false,
 };
 
 MessageContainer.propTypes = {
@@ -204,4 +245,9 @@ MessageContainer.propTypes = {
   inverted: PropTypes.bool,
   loadEarlier: PropTypes.bool,
   invertibleScrollViewProps: PropTypes.object,
+  extraData: PropTypes.object,
+  scrollToBottom: PropTypes.bool,
+  scrollToBottomOffset: PropTypes.number,
+  scrollToBottomComponent: PropTypes.func,
+  alignTop: PropTypes.bool,
 };
