@@ -5,15 +5,19 @@ import React from 'react';
 import { StyleSheet, Text, TouchableOpacity, View, ViewPropTypes } from 'react-native';
 import Color from './Color';
 
-export default function Send({ text, containerStyle, onSend, children, textStyle, label }) {
-  if (text.trim().length > 0) {
+export default function Send({ text, containerStyle, onSend, children, textStyle, label, alwaysShowSend, disabled }) {
+  if (alwaysShowSend || text.trim().length > 0) {
     return (
       <TouchableOpacity
+        testID="send"
+        accessible
+        accessibilityLabel="send"
         style={[styles.container, containerStyle]}
         onPress={() => {
           onSend({ text: text.trim() }, true);
         }}
         accessibilityTraits="button"
+        disabled={disabled}
       >
         <View>{children || <Text style={[styles.text, textStyle]}>{label}</Text>}</View>
       </TouchableOpacity>
@@ -45,6 +49,8 @@ Send.defaultProps = {
   containerStyle: {},
   textStyle: {},
   children: null,
+  alwaysShowSend: false,
+  disabled: false,
 };
 
 Send.propTypes = {
@@ -54,4 +60,6 @@ Send.propTypes = {
   containerStyle: ViewPropTypes.style,
   textStyle: Text.propTypes.style,
   children: PropTypes.element,
+  alwaysShowSend: PropTypes.bool,
+  disabled: PropTypes.bool,
 };
