@@ -1,8 +1,5 @@
 import moment from 'moment';
 
-const DEPRECATION_MESSAGE =
-  'isSameUser and isSameDay should be imported from the utils module instead of using the props functions';
-
 export function isSameDay(currentMessage = {}, diffMessage = {}) {
   if (!diffMessage.createdAt) {
     return false;
@@ -19,17 +16,9 @@ export function isSameDay(currentMessage = {}, diffMessage = {}) {
 }
 
 export function isSameUser(currentMessage = {}, diffMessage = {}) {
-  return !!(
-    diffMessage.user &&
-    currentMessage.user &&
-    diffMessage.user._id === currentMessage.user._id
-  );
+  return !!(diffMessage.user && currentMessage.user && diffMessage.user._id === currentMessage.user._id);
 }
 
-export function warnDeprecated(fn) {
-  return (...args) => {
-    // eslint-disable-next-line
-    console.warn(DEPRECATION_MESSAGE);
-    return fn(...args);
-  };
+export function isExpo() {
+  return !!global && (!!global.__exponent || !!global.__expo);
 }
